backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication
firebase-admin==6.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML Services
groq==0.4.0

# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.8
deepface==0.0.79

# NLP
spacy==3.7.2
textblob==0.17.1

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Audio Processing
SpeechRecognition==3.10.0
pydub==0.25.1

# Report Generation
reportlab==4.0.7

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
pydantic==2.5.0
